feat: §5.5 distortion per-unit need-hearts + parser/engine fixes; bot WIP
Parser (cards/ability_extraction/parser.py):
- 減らし/増やし continuative forms recognized for modify_required_hearts
  operation (was falling to decrease default; 増やす half mis-parsed)
- broadened modify_required_hearts ActionRule matches (増やす/減らす/減らし
  fragments after 、-split)
- _try_per_unit: skip Aし、comma-split when その後 present (dedicated handler
  must win); nested sequential from 。-sentence used directly as fa;
  その後 tail parsed via parse_effect so conditional tails attach properly
- _enrich_heart_content: 必要ハート gate conditions get
  location=live_card_zone + aggregate=total (engine routes to need_heart sum)

Engine:
- score.rs: per-unit counting honors distinct=card_name (duplicate-named
  members count once); self_target candidate pool always includes the
  activating card (live-card abilities resolve wherever they live);
  self_target with no activating card no longer matches all candidates
- condition/card.rs: aggregate-total need_heart sums include
  need_heart_modifiers (gates see effect-modified hearts)
- compound.rs: distinct inherits to ModifyRequiredHearts sub-actions only
  (not gain_resource target selection)
- misc.rs: removed dead reveal→score heuristic (parser now emits proper
  conditional_on_result; heuristic fired even without follow-up text)

Tests: new distortion_need_hearts_test.rs (parse audit + 3/1/dedupe/no-
member cases); qa_new_tests209 hardened to strict outcome asserts.
Regenerated abilities.json + bytecode. 2391 pass, 0 fail.

+++ engine/src/core/game_state/mod.rs
@@ -187,7 +187,7 @@ pub struct GameState {
     /// (e.g. "member card went from live_card_zone to discard this turn").
     pub turn_movements: SmallVec<[MovementEvent; 8]>,
     /// Counter for assigning unique timestamps to MovementEvents within a turn.
-    pub movement_event_counter: u8,
+    pub movement_event_counter: u16,
     /// Snapshot of target cards' orientations taken before a change_state
     /// effect executes. Compared after the effect to detect actual transitions.
     /// None = no snapshot active.
